backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
fastapi==0.104.1
uvicorn==0.24.0

psycopg2-binary # ==2.9.9
python-decouple==3.8
Pillow # ==10.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
requests==2.31.0
